+++ src/grip/index.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+import { Grip } from './Grip'
+
+const StyledGrip = styled(Grip)``
+
+export { GripProps } from './Grip'
+export { StyledGrip as Grip }
+export default StyledGrip