+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
-[...]
+- **[FIX]** Removed margin bottom from `Tabs`
+- **[NEW]** Added `RedCircleIcon`
 
 # v38.0.0 (24/07/2020)
 
@@ -19,6 +20,8 @@
 
 - **[UPDATE]** Add `initialFrom`, `initialTo`, `disabledFrom`, `disabledTo` props in `SearchForm` component
 - **[UPDATE]** Added `secondLine` prop on Tabs config
+
+- **[NEW]** `Grip` component
 
 # v37.0.0 (10/07/2020)